client: Always increment mtime
Cases where the mds clock was a few seconds ahead
of the client would cause the mtime of files to
be set backward, because the mtime set on create
is from the mds, and updates on writes are from the
client.  This patch prevents mtime from going
backward by comparing the current mtime with the
client's time, and incrementing by one nanosecond
if the mtime > client.

Signed-off-by: Sam Lang <sam.lang@inktank.com>