osd: implement prealloc/fallocate object operation
Implement a rados PREALLOC method that will call fallocate(2) to allocate
disk blocks for a while, but not write to them.

We choose the semantics that modify the file size so that the exposed
object metadata will be less confusing.  e.e.g, prealloc to 4MB will result
in a 4MB object full of zeros (or whatever data was prevoiusly written).

Include flags for only doing prealloc on object creation, and for only
doing prealloc on an existing object.

Signed-off-by: Sage Weil <sage@inktank.com>