kvstore: use a standard errno value
Some of the high numbered errno.h values are only available on Linux.

I couldn't find any users in the tree that would be effected by this
change, and it wasn't clear if the value would collide with other errors
or be leaked out to a client.

Are there any users of this library?

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>