Pipe: do not use MSG_MORE on OSX
MSG_MORE is not available on OSX, and unlike MSG_NOSIGNAL, there doesn't
appear to be an equivalent. It is not clear if MSG_MORE effects
correctness or is just an optimization.

If the MSG_MORE semantics are required for correctness, we may be able
to use TCP_CORK to do this on OSX.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>