osd: handle limited osdmap history in project_pg_history
Handle the case where we run out of map history in project_pg_history.
When that happens, assume that the oldest epoch we have is the epoch
where any acting/primary/whatever changed, if we haven't already
identified a later epoch.  This matches what we do with the creation
epoch base case.

Fixes: #5869
Signed-off-by: Sage Weil <sage@inktank.com>