Replace lfn_find in omap* functions with a lightweight implementation
lfn_find plays a important role in handling transactions with FileStore.
Although it's not a heave implementation, it's called several times each
transaction and the cumulative time is still striking.

Introduce a new filestore member to manage object status, it's simply like
a temp container to store whether the object is existing. It need to respond
to the event when object remove happen. And needed to be cleared when
collection split happen.

implement #7063
Signed-off-by: Haomai Wang <haomaiwang@gmail.com>